Remove label link prop from TextBox
This was introduced in https://github.com/desktop/desktop/pull/981 and is only used in one location, 2FA input. As 2FA input in app is exceedingly rare these days with mandatory web flow login we can remove it and at the same time solve a layout issue with very long labels which would previously overflow.